minor code move
