+++ platform-operator/controllers/verrazzano/component/common/http.go
