+++ platform-operator/controllers/verrazzano/component/common/rancher.go
