+++ platform-operator/controllers/verrazzano/component/common/rancher_test.go
