+++ platform-operator/controllers/verrazzano/component/rancher/rancher.go
@@ -6,7 +6,7 @@ package rancher
 import (
 	"fmt"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/common"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/vzconfig"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
